infra/app/requirements.txt
aws_cdk.lib==2.45.0